getting somewhere maybe? but only in macros, they weren't joking.
